Limit time each function can run
So, this was some work I did to try to figure out issue of infinite loops.
The eventual fix for that is in here,
along with code to run the functions in tasks.

Ultimately, I don't think running the low-level functions in tasks is great.
So, I will probably take that out in the next commit,
but I would like to at least have it in the commit history for later reference.