Remove the "task" code for most of ThermoMath
Most of these functions never had issues with running for too long,
and I'm pretty sure I found the fix anyway.
Thus, to get a little more efficiency and less code,
I've removed the "run this in a task" code for almost all the functions.

The only ones still in time-limited tasks are the original offenders,
the "iterate"' functions.
Eventually, I'll probably remove the task code from them too,
but leaving it for the time being.