Add "response power" parameter to dials
Basically, if "response power" is set > 1, we do mapping on val^p,
where p is response power.

That is, when we map to the range of output values for a slider,
given a 0-1 range for the position of the slider, we take val^p.
This gives us a nicer transition on things like weight and balloon,
which can have an absurdly large effect for tiny amounts.