Ensure all classes have at least a little commenting.
Basically, at minimum each class should have a sentence or two at top of file,
which at least tells users what the class/script is used for.
I've done that for all the classes that didn't already have this.
I also expanded on commenting as needed, with the exception of World.cs.
That'll come later, when I've had time to go through that beast of a file.

There is one minor change to code in Touchable,
which should not change any behavior.